fix: 修复安全审计发现的输入校验/整数溢出/并发竞争/内存泄漏漏洞
对电路仿真器做全量安全审计后，系统性修复以下问题（按类别）：

- 拒绝服务：doublebuffer 解码头部无界分配/解压炸弹、Modbus 主站
  响应越界、网表总线展开与节点数无上限、Sync 真值表指数分配
- 整数溢出：unflatten/flatten/TotalRows/矩阵维度乘积/节点号回绕
- 越界：basebyte 长度字段、bitmap 负索引、RLE elemSize/dt=0、
  3D 渲染 NaN、绘图维度乘积、矩阵 row*cols 溢出
- 并发：TimeMNA.elementConverged 原子化、flag 信号丢失、Modbus
  timeout 竞争、TUI droppedUpdates 竞争、事件目标悬垂指针
- 内存泄漏：Animator 完成条目、Modbus 超时读 goroutine
- 协议/规范：Modbus 响应功能码与写线圈值校验、CSR x0 语义、
  MRET/SRET 特权级、C.ADDI4SPN 缩放、SLLI/SRLI/SRAI funct7 校验
- 其他：Sv32 MMU 地址截断、ch34x 动态库加载加固、lu_block NaN
  主元、matrixReducer rowPtr 别名、半导体 Log 参数钳位等

验证：go build ./...、go test ./...、go vet ./...、
go test -race ./element/... ./maths/... 全部通过。

+++ utils/vm/memory.go
@@ -43,7 +43,8 @@ func (dm *DeviceManager) FindDevice(addr uint32) (dev Device, _ uint32) {
 		dev = dm.devices[i]
 		base := dev.GetBaseAddr()
 		size := dev.GetSize()
-		if addr >= base && size > 0 && addr < base+size {
+		// 用 uint64 计算范围，避免 base+size 在 uint32 下回绕导致设备无法命中。
+		if addr >= base && size > 0 && uint64(addr) < uint64(base)+uint64(size) {
 			return dev, addr - base
 		}
 	}
